[ANDROID-31] docs : 온보딩 화면 안에 들어가는 화면 테스트를 패키지 이동
