+++ feature/onboarding/src/androidTest/java/see/day/onboarding/screen/AlertScreenTest.kt
@@ -1,4 +1,4 @@
-package see.day.onboarding
+package see.day.onboarding.screen
 
 import android.content.Context
 import androidx.activity.ComponentActivity
@@ -9,12 +9,11 @@ import androidx.compose.ui.test.onNodeWithText
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
-import see.day.onboarding.screen.OnboardingScreen
-import see.day.onboarding.state.OnboardingScreenState.BIRTHDAY
+import see.day.onboarding.R
+import see.day.onboarding.state.OnboardingScreenState.ALERT
 import see.day.onboarding.state.onboarding.OnboardingUiState
-import java.time.LocalDate
 
-class BirthdayScreenTest {
+class AlertScreenTest {
 
     @get:Rule
     val composeTestRule = createAndroidComposeRule<ComponentActivity>()
@@ -28,7 +27,7 @@ class BirthdayScreenTest {
 
     @Test
     fun givenInitOnboarding_whenScreening_shownRecordTitleAndBody() {
-        val uiState = OnboardingUiState.init.copy(onboardingScreenState = BIRTHDAY)
+        val uiState = OnboardingUiState.init.copy(onboardingScreenState = ALERT)
         composeTestRule.setContent {
             OnboardingScreen(
                 uiState = uiState,
@@ -37,33 +36,11 @@ class BirthdayScreenTest {
         }
 
         composeTestRule
-            .onNodeWithText(context.getString(R.string.birthday_message))
+            .onNodeWithText(context.getString(R.string.alert_message))
             .assertIsDisplayed()
 
         composeTestRule
             .onNodeWithContentDescription("온보딩 ${uiState.onboardingScreenState.ordinal} 번째 아이콘")
             .assertIsDisplayed()
     }
-
-    @Test
-    fun givenEmptyNickname_whenScreening_shownEnableNextButton() {
-        val uiState = OnboardingUiState.init.copy(onboardingScreenState = BIRTHDAY)
-        composeTestRule.setContent {
-            OnboardingScreen(
-                uiState = uiState,
-                uiEvent = {}
-            )
-        }
-
-        composeTestRule
-            .onNodeWithText(LocalDate.now().year.toString())
-            .assertIsDisplayed()
-
-        composeTestRule
-            .onNodeWithText(LocalDate.now().dayOfMonth.toString())
-            .assertIsDisplayed()
-    }
-
-
-
 }
