+++ feature/onboarding/src/androidTest/java/see/day/onboarding/screen/BirthdayScreenTest.kt
@@ -1,4 +1,4 @@
-package see.day.onboarding
+package see.day.onboarding.screen
 
 import android.content.Context
 import androidx.activity.ComponentActivity
@@ -9,12 +9,12 @@ import androidx.compose.ui.test.onNodeWithText
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
-import see.day.onboarding.screen.OnboardingScreen
-import see.day.onboarding.state.OnboardingScreenState.ALERT
+import see.day.onboarding.R
 import see.day.onboarding.state.OnboardingScreenState.BIRTHDAY
 import see.day.onboarding.state.onboarding.OnboardingUiState
+import java.time.LocalDate
 
-class AlertScreenTest {
+class BirthdayScreenTest {
 
     @get:Rule
     val composeTestRule = createAndroidComposeRule<ComponentActivity>()
@@ -28,7 +28,7 @@ class AlertScreenTest {
 
     @Test
     fun givenInitOnboarding_whenScreening_shownRecordTitleAndBody() {
-        val uiState = OnboardingUiState.init.copy(onboardingScreenState = ALERT)
+        val uiState = OnboardingUiState.init.copy(onboardingScreenState = BIRTHDAY)
         composeTestRule.setContent {
             OnboardingScreen(
                 uiState = uiState,
@@ -37,11 +37,33 @@ class AlertScreenTest {
         }
 
         composeTestRule
-            .onNodeWithText(context.getString(R.string.alert_message))
+            .onNodeWithText(context.getString(R.string.birthday_message))
             .assertIsDisplayed()
 
         composeTestRule
             .onNodeWithContentDescription("온보딩 ${uiState.onboardingScreenState.ordinal} 번째 아이콘")
             .assertIsDisplayed()
     }
+
+    @Test
+    fun givenEmptyNickname_whenScreening_shownEnableNextButton() {
+        val uiState = OnboardingUiState.init.copy(onboardingScreenState = BIRTHDAY)
+        composeTestRule.setContent {
+            OnboardingScreen(
+                uiState = uiState,
+                uiEvent = {}
+            )
+        }
+
+        composeTestRule
+            .onNodeWithText(LocalDate.now().year.toString())
+            .assertIsDisplayed()
+
+        composeTestRule
+            .onNodeWithText(LocalDate.now().dayOfMonth.toString())
+            .assertIsDisplayed()
+    }
+
+
+
 }
