+++ feature/onboarding/src/androidTest/java/see/day/onboarding/screen/GoalsScreenTest.kt
@@ -1,4 +1,4 @@
-package see.day.onboarding
+package see.day.onboarding.screen
 
 import android.content.Context
 import androidx.activity.ComponentActivity
@@ -11,8 +11,7 @@ import androidx.compose.ui.test.performClick
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
-import see.day.onboarding.screen.OnboardingScreen
-import see.day.onboarding.state.OnboardingScreenState.BIRTHDAY
+import see.day.onboarding.R
 import see.day.onboarding.state.OnboardingScreenState.GOAL
 import see.day.onboarding.state.onboarding.OnboardingUiState
