+++ feature/onboarding/src/androidTest/java/see/day/onboarding/screen/NicknameScreenTest.kt
@@ -1,4 +1,4 @@
-package see.day.onboarding
+package see.day.onboarding.screen
 
 import android.content.Context
 import androidx.activity.ComponentActivity
@@ -11,7 +11,7 @@ import androidx.compose.ui.test.onNodeWithText
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
-import see.day.onboarding.screen.OnboardingScreen
+import see.day.onboarding.R
 import see.day.onboarding.state.OnboardingScreenState.NICKNAME
 import see.day.onboarding.state.onboarding.OnboardingUiState
